Add files via upload
Your objective is to find a set of three strings containing every permutation of the seven symbols 🎅, 🤶, 🦌, 🧝, 🎄, 🎁, and 🎀 as substrings, subject to the following conditions:

Every permutation must be in at least one string.
Each permutation beginning with 🎅🤶 must be in all three strings.
Each string may have up to two wildcards 🌟, which will match any symbol in a permutation. No string of length seven containing more than one wildcard will count as a permutation.
Your score is the length of the longest of the three strings. This is a minimization problem, so lower scores are better.